# platform/openapi.yaml
openapi: 3.0.3
info:
  title: Platform Service API
  description: |
    User authentication, organization management, invitation system,
    and proxy endpoints to pipeline/plugin microservices.
    Supports JWT and OAuth (Google) authentication.
  version: 0.1.0
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://platform:3000
    description: Internal Docker network
  - url: http://localhost:3000
    description: Local development

# =============================================================================
# Paths
# =============================================================================
paths:
...
  /pipeline:
    get:
      summary: List pipelines
      operationId: listPipelines
      tags: [Pipeline]
      security:
        - bearerAuth: []
      parameters:
        - name: project
          in: query
          schema: { type: string }
        - name: organization
          in: query
          schema: { type: string }
        - name: pipelineName
          in: query
          schema: { type: string }
        - name: isActive
          in: query
          schema: { type: boolean }
        - name: isDefault
          in: query
          schema: { type: boolean }
        - name: accessModifier
          in: query
          schema: { type: string, enum: [public, private] }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20, maximum: 100 }
      responses:
        '200':
          description: Pipelines listed
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      pipelines:
                        type: array
                        items: { type: object }
                      total: { type: integer }
                      page: { type: integer }
                      limit: { type: integer }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }

    post:
      summary: Create a new pipeline
      operationId: createPipeline
      tags: [Pipeline]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [props]
              properties:
                pipelineName: { type: string }
                accessModifier: { type: string, enum: [public, private], default: private }
                props:
                  type: object
                  description: Pipeline builder configuration (project, organization, synth, etc.)
                  required: [project, organization]
                  properties:
                    project: { type: string }
                    organization: { type: string }
      responses:
        '201':
          description: Pipeline created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      pipeline:
                        type: object
                        properties:
                          id: { type: string, format: uuid }
                          project: { type: string }
                          organization: { type: string }
                          pipelineName: { type: string }
                          accessModifier: { type: string }
                          isDefault: { type: boolean }
                          isActive: { type: boolean }
                          createdAt: { type: string, format: date-time }
                          createdBy: { type: string }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /pipeline/{id}:
    get:
      summary: Get pipeline by ID
      operationId: getPipelineById
      tags: [Pipeline]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Pipeline retrieved
          content:
            application/json:
              schema:
                type: object
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }

  # ---------------------------------------------------------------------------
  # Plugin (proxy to plugin service)
  # ---------------------------------------------------------------------------
  /plugin:
    get:
      summary: List plugins
      operationId: listPlugins
      tags: [Plugin]
      security:
        - bearerAuth: []
      parameters:
        - name: name
          in: query
          schema: { type: string }
        - name: version
          in: query
          schema: { type: string }
        - name: pluginType
          in: query
          schema: { type: string }
        - name: computeType
          in: query
          schema: { type: string }
        - name: isActive
          in: query
          schema: { type: boolean }
        - name: isDefault
          in: query
          schema: { type: boolean }
        - name: accessModifier
          in: query
          schema: { type: string, enum: [public, private] }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20, maximum: 100 }
      responses:
        '200':
          description: Plugins listed
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      plugins:
                        type: array
                        items: { type: object }
                      total: { type: integer }
                      page: { type: integer }
                      limit: { type: integer }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }

    post:
      summary: Upload and create a new plugin
      operationId: createPlugin
      tags: [Plugin]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [plugin]
              properties:
                plugin:
                  type: string
                  format: binary
                  description: ZIP archive containing spec and Dockerfile
                accessModifier:
                  type: string
                  enum: [public, private]
                  default: private
      responses:
        '201':
          description: Plugin created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      plugin:
                        type: object
                        properties:
                          id: { type: string, format: uuid }
                          name: { type: string }
                          version: { type: string }
                          uri: { type: string, description: 'Computed image URI: <namespace>/<name>:<version>' }
                          fullImage: { type: string }
                          accessModifier: { type: string }
                          isDefault: { type: boolean }
                          isActive: { type: boolean }
                          createdBy: { type: string }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /plugin/{id}:
    get:
      summary: Get plugin by ID
      operationId: getPluginById
      tags: [Plugin]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Plugin retrieved
          content:
            application/json:
              schema:
                type: object
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }

  # ---------------------------------------------------------------------------
  # Admin
  # ---------------------------------------------------------------------------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token
    refreshToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT refresh token from login response
...
  schemas:
    SuccessResponse:
      type: object
      required: [success, statusCode]
      properties:
        success: { type: boolean, enum: [true] }
        statusCode: { type: integer }
...
    ErrorResponse:
      type: object
      required: [success, statusCode, message]
      properties:
        success: { type: boolean, enum: [false] }
        statusCode: { type: integer }
        message: { type: string }
        code: { type: string }
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
security:
  - bearerAuth: []
...